refactor(services): extract focused modules to satisfy max-lines without suppressions
- background-pane-env.ts: one shared background pane env builder for the
  terminal and agent-session launchers (was duplicated shape).
- worktree-service-provisioning-flow.ts: provision event stream + post-
  create hydration out of worktree-creation-flow.
- worktree-startup-payload.ts: pure buildStartupOpt out of
  worktree-creation-flow (file exceeded max-lines after rebase onto main).
