+++ src/renderer/src/lib/launch-agent-background-session.ts
@@ -25,7 +25,7 @@ import {
 import { subscribeToPtyData } from '@/components/terminal-pane/pty-data-sidecar-subscriptions'
 import { callRuntimeRpc, getActiveRuntimeTarget } from '@/runtime/runtime-rpc-client'
 import { getSettingsForWorktreeRuntimeOwner } from '@/lib/worktree-runtime-owner'
-import { getWorktreeServiceEnv } from '@/lib/worktree-service-env-injection'
+import { buildBackgroundPaneEnv } from '@/lib/background-pane-env'
 import { toRuntimeWorktreeSelector } from '@/runtime/runtime-worktree-selector'
 import { singlePaneLayoutSnapshot } from '@/store/slices/terminal-helpers'
 import { retireProvider, retireUnownedTerminal } from '@/lib/retire-unowned-background-terminal'
@@ -127,11 +127,7 @@ export async function launchAgentBackgroundSession(
   // in-terminal title row, so background sessions must not persist it there.
   store.setTabLayout(tab.id, singlePaneLayoutSnapshot(leafId))
   const paneEnv = {
-    ...startupPlan.env,
-    ...getWorktreeServiceEnv(worktreeId),
-    ORCA_PANE_KEY: paneKey,
-    ORCA_TAB_ID: tab.id,
-    ORCA_WORKTREE_ID: worktreeId,
+    ...buildBackgroundPaneEnv(worktreeId, tab.id, leafId, startupPlan.env),
     ORCA_AGENT_LAUNCH_TOKEN: launchToken
   }
   const sshConnectionId = repo?.connectionId ?? null
