+++ src/renderer/src/lib/worktree-startup-payload.ts
@@ -1,0 +1,30 @@
+import type { WorktreeStartupPayload } from '@/lib/worktree-activation'
+import type { WorktreeCreationRequest } from '@/lib/pending-worktree-creation'
+
+// Why: when the backend already spawned the startup terminal, sending a second
+// startup payload would double-launch the agent; the payload only exists for
+// the renderer-driven fallback.
+export function buildStartupOpt(
+  request: WorktreeCreationRequest,
+  backendSpawned: boolean
+): WorktreeStartupPayload | undefined {
+  const plan = request.startupPlan
+  if (!plan || backendSpawned) {
+    return undefined
+  }
+  return {
+    command: plan.launchCommand,
+    ...(plan.env ? { env: plan.env } : {}),
+    launchConfig: plan.launchConfig,
+    ...(plan.launchToken ? { launchToken: plan.launchToken } : {}),
+    ...(request.agent ? { launchAgent: request.agent } : {}),
+    ...(plan.draftPrompt ? { draftPrompt: plan.draftPrompt } : {}),
+    ...(plan.startupCommandDelivery ? { startupCommandDelivery: plan.startupCommandDelivery } : {}),
+    // Why: command-code shows its prompt in the tab status before the first
+    // hook fires, so the prompt is threaded through here.
+    ...(request.agent === 'command-code' && request.quickPrompt.trim().length > 0
+      ? { initialAgentStatus: { agent: request.agent, prompt: request.quickPrompt.trim() } }
+      : {}),
+    ...(request.quickTelemetry ? { telemetry: request.quickTelemetry } : {})
+  }
+}
